+++ src/canvas.rs
@@ -66,14 +66,21 @@ impl Canvas {
         self.size
     }
 
-    /// Set value of \[`row`, `col`\] to `cell`.
+    /// Sets the cell at (`row`, `col`) to the value `cell`.
     pub fn set_cell(&mut self, row: u32, col: u32, cell: Cell) {
         debug_assert!(row < self.size.rows);
         debug_assert!(col < self.size.cols);
-
         self.back.set_cell(row, col, cell);
     }
 
+    /// Sets the cell at (`row`, `col`) to the value `c` using `color`.
+    #[inline]
+    pub fn set(&mut self, row: u32, col: u32, c: char, color: Color) {
+        self.set_cell(row, col, Cell::new(c, color));
+    }
+
+    /// Writes `c` using `color` at (`row`, `col`), returning `1` if the character
+    /// was written or `0` if `col` extends beyond the end of the row.
     pub fn write_char(&mut self, row: u32, col: u32, c: char, color: Color) -> u32 {
         if col < self.size.cols {
             self.back.set_cell(row, col, Cell::new(c, color));
@@ -83,11 +90,11 @@ impl Canvas {
         }
     }
 
-    pub fn write_str(&mut self, row: u32, col: u32, chars: &str, color: Color) -> u32 {
-        let chars = chars.chars().collect::<Vec<_>>();
-        self.write(row, col, &chars, color)
-    }
-
+    /// Write `chars` using `color` in `row` starting at column `col`, returning the
+    /// actual number of characters written.
+    ///
+    /// Fewer than `chars.len()` may be written if doing so would extend beyond the
+    /// end of the row.
     pub fn write(&mut self, row: u32, col: u32, chars: &[char], color: Color) -> u32 {
         debug_assert!(row < self.size.rows);
         let n = cmp::min(chars.len(), (self.size.cols - col) as usize);
@@ -98,8 +105,14 @@ impl Canvas {
         n as u32
     }
 
-    /// Fills all cells with `cell` in the column range [`start_col`, `end_col`)
-    /// for the given `row`.
+    /// Write `chars` using `color` in `row` starting at column `col`, returning the
+    /// actual number of characters written.
+    pub fn write_str(&mut self, row: u32, col: u32, chars: &str, color: Color) -> u32 {
+        let chars = chars.chars().collect::<Vec<_>>();
+        self.write(row, col, &chars, color)
+    }
+
+    /// Fills cells in `row` for the column range `cols` with the value `cell`.
     pub fn fill_cell(&mut self, row: u32, cols: Range<u32>, cell: Cell) {
         debug_assert!(row < self.size.rows);
         let Range { start, end } = cols;
@@ -108,16 +121,28 @@ impl Canvas {
         self.back.fill_range(row, start, end, cell);
     }
 
+    /// Fills cells in `row` for the column range `cols` with the value `c` using
+    /// `color`.
+    #[inline]
     pub fn fill(&mut self, row: u32, cols: Range<u32>, c: char, color: Color) {
         self.fill_cell(row, cols, Cell::new(c, color));
     }
 
-    /// Fills all cells with `cell` in the column range [`start_col`, ..) for the
-    /// given `row`.
+    /// Fills cells in `row` starting at column `start_col` with the value `cell`.
+    #[inline]
     pub fn fill_cell_from(&mut self, row: u32, start_col: u32, cell: Cell) {
         self.fill_cell(row, start_col..self.size.cols, cell);
     }
 
+    /// Fills cells in `row` starting at column `start_col` with the value `c` using
+    /// `color`.
+    #[inline]
+    pub fn fill_from(&mut self, row: u32, start_col: u32, c: char, color: Color) {
+        self.fill_cell_from(row, start_col, Cell::new(c, color));
+    }
+
+    /// Fills all cells in `row` with the value `c` using `color`.
+    #[inline]
     pub fn fill_row(&mut self, row: u32, c: char, color: Color) {
         self.fill(row, 0..self.size.cols, c, color);
     }
@@ -126,7 +151,6 @@ impl Canvas {
     pub fn set_cursor(&mut self, cursor: Point) {
         debug_assert!(cursor.row < self.size.rows);
         debug_assert!(cursor.col < self.size.cols);
-
         self.writer.set_cursor(cursor);
         self.writer.send();
     }