+++ src/echo.rs
@@ -1,17 +1,23 @@
 //! Echo messages to the shared region of a workspace.
 
+use crate::color::Color;
 use crate::workspace::WorkspaceRef;
 use crate::writer::Writer;
 
 pub struct Echo {
     workspace: WorkspaceRef,
+    echo_color: Color,
     text: Option<String>,
 }
 
 impl Echo {
     pub fn new(workspace: WorkspaceRef) -> Echo {
+        let config = workspace.borrow().config().clone();
+        let echo_color = Color::new(config.theme.echo_fg, config.theme.text_bg);
+
         Echo {
             workspace,
+            echo_color,
             text: None,
         }
     }
@@ -35,7 +41,7 @@ impl Echo {
             let blank_cols = size.cols - chars.len() as u32;
 
             Writer::new_at(origin)
-                .set_color(self.workspace.borrow().config().theme.echo_color)
+                .set_color(self.echo_color)
                 .write_str(chars.into_iter().collect::<String>().as_str())
                 .write_str(" ".repeat(blank_cols as usize).as_str())
                 .send();